STL porting
git-svn-id: svn://svn.savannah.gnu.org/myserver/trunk@1219 c4c922ea-a70d-0410-937d-ec1daf91936c

+++ myserverweb/source/http_file.cpp
@@ -279,7 +279,7 @@ int HttpFile::send(HttpThreadContext* td, ConnectionPtr s, char *filenamePath,
 					{
             ostringstream buffer;
 						h.closeFile();
-            buffer << (int)dataSent;
+            buffer << dataSent;
             td->response.CONTENT_LENGTH.assign(buffer.str().c_str());
 						return 0;
 					}
